VMS-15751: English in comments fix
VMS-16284: Review fixes
VMS-16284: Use bypass for advanced PTZ if possible
VMS-16975: mediaserver ec2/analyticsLookupObjectTracks request hanged
VMS-17150: Added information about changes in /ec2/cameraThumbnail to the API changelog
VMS-17150: Fixed an analytics db UT
VMS-17150: Fixed minor issues, added unit tests
VMS-17150: Post-merge analytics settings QML fixes
VMS-17150: Removed sameAsAnalytics option for the cameraThumbnail request
VMS-17274: Added UI TimeZone setting.
VMS-17418: rtsp_perf: Refactor: Remove not needed inheritance from Singleton mixin
VMS-17537: Fix UT build
VMS-17537: Fix review issues: remove ServerModule dependency from PlatformAbstraction
VMS-17562: Stub analytics plugin: added a separate group box for the plugin-side polygon
VMS-17596: Run dependency handler chains on updating parameters visibility
VMS-17609: 'Previous Frame' doesn't work on border Hi-Res <-> Low-Res
VMS-17611: VMS Metrics: fix assert about offlineFromOnlineStatus
VMS-17618: The VIVOTEK SC8131 camera is not autodiscovered
VMS-17623: Resolve smb host name before mount
VMS-17630: Do not send data in QnRtspDataConsumer if it is stopping
Branch: vms_4.1

+++ vms/libs/nx_sdk/src/nx/sdk/helpers/media_stream_statistics.cpp
@@ -60,7 +60,7 @@ void MediaStreamStatistics::onData(
     removeRange(toIterator(timestamp + m_windowSize), m_data.end());
     if (!m_data.empty())
         removeRange(m_data.begin(), toIterator(m_data.back().timestamp - m_windowSize));
-    const int extraFrames = m_data.size() - m_maxDurationInFrames;
+    const int extraFrames = (int) m_data.size() - m_maxDurationInFrames;
     if (extraFrames > 0 && m_maxDurationInFrames > 0)
         removeRange(m_data.begin(), m_data.begin() + extraFrames);
     m_lastDataTimer = steady_clock::now();
